tests/wpt/tests/css/css-ruby/parsing/ruby-position-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Ruby Layout: parsing ruby-position with valid values</title>
<link rel="help" href="https://drafts.csswg.org/css-ruby-1/#propdef-ruby-position">
<meta name="assert" content="ruby-position supports the full grammar 'over | under | inter-character'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("ruby-position", "over");
test_valid_value("ruby-position", "under");
test_valid_value("ruby-position", "inter-character");
</script>
</body>
</html>
